+++ src/Dto/Customer/Customer/Dto/Attribute.php
@@ -27,18 +27,17 @@ use Shopgate\ConnectSdk\Dto\Customer\Customer\Dto\Attribute\Value;
 /**
  * Dto for customer attribute.
  *
- * @method setCode(string $code)
- * @method setValue(string|Value $value)
- * @method setName(string $name)
+ * @method Attribute setCode(string $code)
+ * @method Attribute setName(string $name)
  * @method string getCode()
- * @method string|Value getValue()
  * @method string getName()
+ *
+ * @codeCoverageIgnore
  */
 class Attribute extends Base
 {
     /**
      * @var array
-     * @codeCoverageIgnore
      */
     protected $schema = [
         'type'                 => 'object',
@@ -49,4 +48,31 @@ class Attribute extends Base
         ],
         'additionalProperties' => true,
     ];
+
+    /**
+     * @param Value|string $value
+     *
+     * @return Attribute
+     */
+    public function setValue($value)
+    {
+        if (is_object($value)) {
+            return $this::set('value', new Value($value));
+        }
+
+        return $this::set('value', $value);
+    }
+
+    /**
+     * @return Value|string|null
+     */
+    public function getValue()
+    {
+        $value = $this::get('value');
+        if (is_object($value)) {
+            return new Value($value);
+        }
+
+        return $value;
+    }
 }