+++ tests/Integration/Dto/Customer/CustomerTest.php
@@ -246,11 +246,10 @@ class CustomerTest extends CustomerBaseTest
 
         $attribute = new Customer\Dto\Attribute();
         $attribute->setCode(self::CUSTOMER_ATTRIBUTE_CODE);
-        $attribute->setName('Main color');
-
+        $attribute->setName('Attribute Name');
         $value = new Customer\Dto\Attribute\Value();
         $value->setCode(self::CUSTOMER_ATTRIBUTE_VALUE_CODE);
-
+        $value->setName('Main color');
         $attribute->setValue($value);
 
         // Add attribute
@@ -316,13 +315,13 @@ class CustomerTest extends CustomerBaseTest
         $this->assertEquals(self::CUSTOMER_SETTINGS_DEFAULT_LOCALE, $settings->getDefaultLocale());
         $this->assertEquals(self::CUSTOMER_SETTINGS_DEFAULT_CURRENCY, $settings->getDefaultCurrency());
         $this->assertEquals(self::CUSTOMER_SETTINGS_DEFAULT_LOCATION_CODE, $settings->getDefaultLocationCode());
-        $this->assertCount(2, $settings->getCommunicationPreferences()->toArray());
-        $this->assertTrue(in_array('email', $settings->getCommunicationPreferences()->toArray()));
-        $this->assertTrue(in_array('sms', $settings->getCommunicationPreferences()->toArray()));
+        $this->assertCount(2, $settings->getCommunicationPreferences());
+        $this->assertContains('email', $settings->getCommunicationPreferences()->toArray());
+        $this->assertContains('sms', $settings->getCommunicationPreferences()->toArray());
         $this->assertEquals(true, $settings->getMarketingOptIn());
 
         //Contacts
-        $this->assertCount(1, $contacts->toArray());
+        $this->assertCount(1, $contacts);
         $contact = $contacts[0];
         $this->assertEquals(self::CUSTOMER_CONTACT_EXTERNAL_CUSTOMER_CODE, $contact->getExternalContactCode());
         $this->assertEquals(self::CUSTOMER_CONTACT_FIRSTNAME, $contact->getFirstName());
@@ -345,7 +344,7 @@ class CustomerTest extends CustomerBaseTest
         $this->assertEquals(true, $contact->getIsDefaultShipping());
 
         //Attributes
-        $this->assertCount(1, $attributes->toArray());
+        $this->assertCount(1, $attributes);
         $attribute = $attributes[0];
 
         $this->assertEquals(self::CUSTOMER_ATTRIBUTE_NAME, $attribute->getName());
@@ -435,9 +434,9 @@ class CustomerTest extends CustomerBaseTest
         $this->assertEquals('de-de', $settings->getDefaultLocale());
         $this->assertEquals('USD', $settings->getDefaultCurrency());
         $this->assertEquals('EUR', $settings->getDefaultLocationCode());
-        $this->assertCount(2, $settings->getCommunicationPreferences()->toArray());
-        $this->assertTrue(in_array('sms', $settings->getCommunicationPreferences()->toArray()));
-        $this->assertTrue(in_array('email', $settings->getCommunicationPreferences()->toArray()));
+        $this->assertCount(2, $settings->getCommunicationPreferences());
+        $this->assertContains('sms', $settings->getCommunicationPreferences()->toArray());
+        $this->assertContains('email', $settings->getCommunicationPreferences()->toArray());
         $this->assertEquals(false, $settings->getMarketingOptIn());
     }